+++ examples/definitions.ts
@@ -16,14 +16,17 @@
  * limitations under the License.
  */
 
+// Run this sample with: npx ts-node examples/definitions.ts
+
 import { Reader, ResultType } from "../src/reader";
 import { Tracker, MessageType } from "../src/tracker";
+import * as path from "path";
 
 import * as fs from "fs";
 
 let document: string = "";
-document += "UNB+UNOA:1:::::::::::::::::::::+005435656:1+006415160:1+060515:1434+00000000000778'";
-document += "UNH+00000000000117+INV\n\rOIC:D:97B:UN'";
+document += "UNB+UNOA:1+005435656:1+006415160:1+060515:1434+00000000000778'";
+document += "UNH+00000000000117+INV\n\rOIC:D:01B:UN'";
 document += "BGM+380+342459+9'";
 document += "DTM+3:20060515:102'";
 document += "RFF+ON:521052'";
@@ -54,7 +57,7 @@ function _validateDocument(doc: string, callback?: (numChecked: number) => void)
     const result: ResultType[] = reader.parse(doc);
 
     let checked: number = 0;
-    const data: string = fs.readFileSync("INVOIC.json", { encoding: "utf-8"});
+    const data: string = fs.readFileSync(path.resolve("./src/messageSpec/INVOIC.struct.json"), { encoding: "utf-8"});
     const msgStruct: MessageType[] = JSON.parse(data) as MessageType[];
     const tracker: Tracker = new Tracker(msgStruct);