+++ spec/reader.spec.ts
@@ -26,7 +26,7 @@ describe("Edifact Reader", () => {
         // custom decimal character in UNA segment!
         document += "UNA:+,? '";
         document += "UNB+UNOA:1+005435656:1+006415160:1+060515:1434+00000000000778'";
-        document += "UNH+00000000000117+INV\n\rOIC:D:97B:UN'";
+        document += "UNH+00000000000117+INV\n\rOIC:D:01B:UN'";
         document += "BGM+380+342459+9'";
         document += "DTM+3:20060515:102'";
         document += "RFF+ON:521052'";
@@ -55,7 +55,7 @@ describe("Edifact Reader", () => {
 
     it("should read document", () => {
 
-        const sut: Reader = new Reader();
+        const sut: Reader = new Reader("./src/messageSpec");
 
         const parsingResult: ResultType[] = sut.parse(document);